Move: Component폴더 속 파일 shared로 이동
