+++ portfolio-project/src/app/@modal/(.)feedback/page.tsx
@@ -1,4 +1,4 @@
-import FeedbackContent from '@/components/FeedbackContent';
+import FeedbackContent from '@/shared/ui/FeedbackContent';
 
 export default function FeedbackModal() {
   return <FeedbackContent isModal={true} />;
