+++ portfolio-project/src/widgets/blog-section.tsx
@@ -1,4 +1,4 @@
-'use Client';
+'use client';
 import Image from 'next/image';
 
 export default function BlogSection() {
